написал логику эндпоинтов для admin

+++ src/main/java/com/example/prodBack8/model/entity/group/GroupEntity.java
@@ -17,7 +17,7 @@ public class GroupEntity extends BaseEntity {
 
     private String name;
     private Integer GPUcount;
-    private Double distribution;
+    private Integer distribution;
 
     @Embedded
     private UsageLimit usageLimit;
